+++ core/ledger/kvledger/txmgmt/statedb/statecouchdb/couchdb_test_export.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/hyperledger/fabric/common/metrics/disabled"
 	"github.com/hyperledger/fabric/core/ledger"
-	"github.com/hyperledger/fabric/integration/runner"
+	"github.com/hyperledger/fabric/integration/nwo/runner"
 	"github.com/stretchr/testify/require"
 )
